Ensure pods are deleted with remote Job by setting delete propagation policy (#5607)
In a MultiKueue setup, deleting a Job in the manager cluster does not remove pods in the worker cluster, leading to orphaned pods that continue consuming resources.

This change adds `PropagationPolicy: Background` to the `remoteClient.Delete()` call in the MultiKueue adapter to ensure the remote Job's pods are cleaned up properly.

Signed-off-by: ichekrygin <illya.chekrygin@gmail.com>
Co-authored-by: ichekrygin <illya.chekrygin@gmail.com>